third_party/WebKit/LayoutTests/dom/node/isSupported.html

<!doctype html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Node.isSupported should not exist");
var node = new DocumentFragment();
shouldBeFalse("'isSupported' in node");
</script>
</body>
</html>
