third_party/WebKit/LayoutTests/dom/node/isSupported-expected.txt

Node.isSupported should not exist

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 'isSupported' in node is false
PASS successfullyParsed is true

TEST COMPLETE
